Publish system events on space prefixed path for hosted version
