+++ .travis.yml
@@ -14,6 +14,7 @@ script:
 - gometalinter --install --force
 - gometalinter --vendor --fast --disable=gotype --disable=vetshadow --disable=gas --skip=mock ./...
 - ./codecov.sh
+- go test ./router -tags=hosted
 - go test ./tests -tags=integration
 after_success:
 - test -n "$TRAVIS_TAG" && curl -sL https://git.io/goreleaser | bash
