+++ router/router_test.go
@@ -61,20 +61,6 @@ func TestRouterServeHTTP(t *testing.T) {
 		assert.Equal(t, "http://example.com", recorder.Header().Get("Access-Control-Allow-Origin"))
 	})
 
-	t.Run("extract path from hosted domain", func(t *testing.T) {
-		target.EXPECT().CORS(gomock.Any(), gomock.Any()).Return(nil)
-		target.EXPECT().SyncSubscriber(http.MethodGet, "/custom/test", event.TypeName("http.request")).Return(nil).MaxTimes(1)
-		target.EXPECT().AsyncSubscribers(http.MethodGet, "/custom/test", event.TypeName("http.request")).Return([]router.AsyncSubscriber{}).MaxTimes(1)
-		target.EXPECT().AsyncSubscribers(http.MethodPost, "/", event.SystemEventReceivedType).Return([]router.AsyncSubscriber{}).MaxTimes(1)
-		router := setupTestRouter(target)
-
-		req, _ := http.NewRequest(http.MethodGet, "https://custom.slsgateway.com/test", nil)
-		recorder := httptest.NewRecorder()
-		router.ServeHTTP(recorder, req)
-
-		assert.Equal(t, http.StatusAccepted, recorder.Code)
-	})
-
 	t.Run("reject if system event", func(t *testing.T) {
 		target.EXPECT().CORS(gomock.Any(), gomock.Any()).Return(nil)
 		router := setupTestRouter(target)
